Moved 2 event functions out of graphclass and added constant
